+++ reactionmenu/text.py
@@ -96,15 +96,23 @@ class TextMenu(abc.Menu):
                 Added :attr:_menu_timed_out
             v2.0.1
 				Added :attr:`_bypass_primed`
+            v2.0.3
+				Added instantiation of :attr:`Button.menu`
     """
     _active_sessions: List['TextMenu'] = []
 
     def __init__(self, ctx: Context, *, back_button: str, next_button: str, **options):
         self._ctx = ctx
         self._bot = ctx.bot
-        self._default_back_button: Button = Button(emoji=back_button, linked_to=ButtonType.PREVIOUS_PAGE, name='default back button')
-        self._default_next_button: Button = Button(emoji=next_button, linked_to=ButtonType.NEXT_PAGE, name='default next button')
+        
+        back_obj = Button(emoji=back_button, linked_to=ButtonType.PREVIOUS_PAGE, name='default back button')
+        next_obj = Button(emoji=next_button, linked_to=ButtonType.NEXT_PAGE, name='default next button')
+        back_obj._Button__menu = self
+        next_obj._Button__menu = self
+        self._default_back_button: Button = back_obj
+        self._default_next_button: Button = next_obj
         self._all_buttons: List[Button] = [self._default_back_button, self._default_next_button]
+        
         self._contents: List[str] = []
         self._loop = ctx.bot.loop
         self._send_to_channel = None
@@ -247,6 +255,10 @@ class TextMenu(abc.Menu):
         
             .. note::
                 ABC meth
+            
+            .. changes::
+				v2.0.3
+					Addded instantiation of :attr:`Button.menu`
         """
         if button.emoji not in self._extract_all_emojis():
             
@@ -280,6 +292,8 @@ class TextMenu(abc.Menu):
 
             if self._all_buttons_removed:
                 self._all_buttons_removed = False
+            
+            button._Button__menu = self
         else:
             raise DuplicateButton(f'The emoji {tuple(button.emoji)} has already been registered as a button')
 
@@ -403,6 +417,7 @@ class TextMenu(abc.Menu):
                     # back
                     if emoji == btn.emoji and btn.linked_to is ButtonType.PREVIOUS_PAGE:
                         worker = self._contents[self._page_control('-')]
+                        self._update_button_statistics(btn, user)
                         await self._execute_navigation_type(worker, btn.emoji)
                         await self._contact_relay(user, btn)
                         break
@@ -410,6 +425,7 @@ class TextMenu(abc.Menu):
                     # next
                     elif emoji == btn.emoji and btn.linked_to is ButtonType.NEXT_PAGE:
                         worker = self._contents[self._page_control('+')]
+                        self._update_button_statistics(btn, user)
                         await self._execute_navigation_type(worker, btn.emoji)
                         await self._contact_relay(user, btn)
                         break
@@ -418,6 +434,7 @@ class TextMenu(abc.Menu):
                     elif emoji == btn.emoji and btn.linked_to is ButtonType.GO_TO_FIRST_PAGE:
                         self._page_number = 0
                         worker = self._contents[self._page_number]
+                        self._update_button_statistics(btn, user)
                         await self._execute_navigation_type(worker, btn.emoji)
                         await self._contact_relay(user, btn)
                         break
@@ -426,6 +443,7 @@ class TextMenu(abc.Menu):
                     elif emoji == btn.emoji and btn.linked_to is ButtonType.GO_TO_LAST_PAGE:
                         self._page_number = len(self._contents) - 1
                         worker = self._contents[self._page_number]
+                        self._update_button_statistics(btn, user)
                         await self._execute_navigation_type(worker, btn.emoji)
                         await self._contact_relay(user, btn)
                         break
@@ -465,6 +483,7 @@ class TextMenu(abc.Menu):
                                 if requested_page >= 1 and requested_page <= self.total_pages:
                                     self._page_number = requested_page - 1
                                     worker = self._contents[self._page_number]
+                                    self._update_button_statistics(btn, user)
                                     await self._execute_navigation_type(worker, btn.emoji)
                                     await self._contact_relay(user, btn)
                                     if self._delete_interactions:
@@ -489,7 +508,8 @@ class TextMenu(abc.Menu):
                             except TypeError as invalid_args:
                                 raise ReactionMenuException(f'{ERROR_MESSAGE}: {invalid_args}')
 
-                        # worker param is :class:`None` just as a placeholder. It is not handled in the call			
+                        # worker param is :class:`None` just as a placeholder. It is not handled in the call
+                        self._update_button_statistics(btn, user)
                         await self._execute_navigation_type(None, btn.emoji, from_caller_button=True)
                         await self._contact_relay(user, btn)
                         break